+++ src/provider/DummyProvider.php
@@ -2,19 +2,20 @@
 
 namespace alexeevdv\sms\provider;
 
+use alexeevdv\Sms\Contract\Provider;
 use yii\base\BaseObject;
 
 /**
  * Class DummyProvider
  * @package alexeevdv\sms\provider
  */
-final class DummyProvider extends BaseObject implements ProviderInterface
+final class DummyProvider extends BaseObject implements Provider
 {
     /**
      * @inheritdoc
      */
-    public function send($number, $text)
+    public function sendMessage($number, $text)
     {
-        return true;
+        return new MessageId('');
     }
 }